Title: 합과 차의 나눗셈, Time: -1ms, Memory: -1MB, Status: Runtime Error - CodeTree

+++ 240509/합과 차의 나눗셈/divide-of-sum-and-sub.py
@@ -1,1 +1,1 @@
-a,b=map(int,input().split());print(f"{(a+b)/(a-b):.2f}"))
+a,b=map(float,input().split());print(f"{(a+b)/(a-b):.2f}"))
